requirements.txt
crewai>=0.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
